Add a project poster template
This changeset adds a template to use for project
posters. This is adapted from "corporate world," so
happy to make further changes that would make it more
appropriate for the kind of work we do.

I've also included an example by trying to backfill a
project poster for Cladetime. It's not a great example,
to be honst--it's hard to write these things retrospectively.